requirements.txt
requests==2.31.0
python-dotenv==1.0.0
discord.py==2.3.2
python-telegram-bot==20.7
schedule==1.2.1
flask==3.0.0
stripe==8.0.0
werkzeug==3.0.1
